requirements.txt
absl-py==0.10.0
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
blinker==1.4
boto3==1.21.3
botocore==1.24.3
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
cycler==0.11.0
Django==4.0.2
django-ninja==0.17.0
django-cors-headers==3.11.0
flatbuffers==2.0
fonttools==4.29.1
frozenlist==1.2.0
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.1.0
idna==3.3
importlib-metadata==4.10.1
jmespath==0.10.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.1
multidict==5.2.0
munkres==1.1.4
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
Pillow==9.0.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
pyu2f==0.1.5
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
s3transfer==0.5.1
scipy==1.7.3
six==1.15.0
sqlparse==0.4.2
tensorboard==2.6.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.7.0
tensorflow==2.7.0
tensorflow-hub==0.12.0
termcolor==1.1.0
tornado==6.1
typing-extensions==3.7.4.3
unicodedata2==14.0.0
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.12.1
yarl==1.7.2
zipp==3.7.0